requirements.txt
# MarketScholar Requirements
# Python 3.11+

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.9  # Backup for direct PostgreSQL connection

# Market data
yfinance>=0.2.36

# Gemini AI with Grounding
google-generativeai>=0.8.0

# Data analysis
numpy>=1.26.0
pandas>=2.2.0

# Optional but recommended
python-dotenv>=1.0.0  # For local development

# Testing (optional)
pytest>=8.0.0
pytest-cov>=4.1.0
